Add GopherJS compatibility build tags
This adds build tags to exclude the unix-specific stuff from a js build.
