+++ terminal_linux.go
@@ -4,6 +4,7 @@
 // license that can be found in the LICENSE file.
 
 // +build !appengine
+// +build !js
 
 package logrus
