+++ text_formatter_js.go
@@ -1,0 +1,11 @@
+// +build js
+
+package logrus
+
+import (
+	"io"
+)
+
+func (f *TextFormatter) checkIfTerminal(w io.Writer) bool {
+	return false
+}
